requirements.txt
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
paypalrestsdk==1.13.3
requests==2.32.3
psycopg2-binary==2.9.9
whitenoise==6.7.0
dj-database-url==2.2.0
gunicorn==22.0.0
